openforis/sepal#18

+++ sepal-server/src/main/groovy/org/openforis/sepal/metadata/crawling/MetadataCrawlingCriteria.groovy
@@ -6,7 +6,7 @@ import groovy.transform.Immutable
 @Immutable
 class MetadataCrawlingCriteria {
 
-    int criteriaId
+    Long criteriaId
     String fieldName
     String expectedValue
 }
